Isolate every workpieces from image
To have a better analysis for every workpieces, need to isolate them from image.
Therefore, using mask image (*.bmp) to get the single mask image of every workpiece.

On the other hand, there are some pictures that are difficult to get contours properly,
so update the image to get more better performance.